Enable actions to be submitted via a POST request
To enable the dewar handling barcode scanner system to update the status
of dewars when they arrrive and depart we add an endpoint that actions
can be posted to.

The handleAction function is expected to return a promise but was
returning undefined in the case of an unrecognised action type.  We now
return a rejected promise instead.