Update receptacle for pucks in shipped dewars
When a user submits a dewar shipment form the pucks they say they are
shipping should be added to the dewar. This is complicated by the
possibility the pucks they are shipping are not in the system. In this
change we do our best at parsing the user input and if there are
matching pucks that are not already in a receptacle we set their
location to the newly added dewar.