Remove dewar info from pucks when dewar is deleted
Prior to this change, when a dewar was deleted from the system pucks
could remain labelled as being inside this dewar. This could cause
confusion and prevent the system correctly updating puck locations when
users shipped a new dewar. This change fixes this by clearing the pucks'
receptacle information when the dewar is deleted.